requirements.txt
streamlit
pandas
openai
python-dotenv
duckduckgo-search
graphviz
plotly
